feat(contrib/ec2): add setup-proxy-protocol.sh
setup-proxy-protocol.sh makes it quick and easy to enable proxy protocol
support on your AWS CoreOS/Deis cluster.

It is a script since it also needs to set a property through deisctl to
enable Proxy Protocol support in deis-router. If the value is not set
then the router will not accept incoming requests from Amazon ELB.