docs(websocket): Add WebSocket setup documentation
The documentation contains information on how to setup
WebSockets on Deis.

Deis have had support for WebSockets for a while, but
there was little documentation on the steps necessary to
make it work in the cloud.

The provided documentation adds information on setting
up WebSockets on AWS; Rackspace & Digital Ocean.