docs(websocket): Add WebSocket setup documentation
The documentation contains information on how to setup
WebSockets on Deis.

Deis have had support for WebSockets for a while, but
there was little documentation on the steps necessary to
make it work in the cloud.

The provided documentation adds information on setting
up WebSockets on AWS; Rackspace & Digital Ocean.

The AWS Instructions also include information on getting
it running with the PROXY protocol, which allows the
server to know the requests original  IP information.

Rackspace does not support the Proxy protocol.

DigitalOcean doesn't need it (since it uses DNS based
load balancing).