chore(contrib/ec2): Deprecate Outputs:DNSName
Deprecate the Outputs:DNSName. Since we now have the possibility of
having two load balancers, it is favourable to deprecate DNSName in
favour of WebELBDNSName and TCPELBDNSName, which is more clearer.

In case anyone were using the DNSName output (I was, so others might
be as well), it is deprecated instead of removed, giving users time
to upgrade.